Remove AppContext.CoreCLR.cs (move to shared)
